requirements.txt
Django==3.2.16
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
django-environ==0.8.1
psycopg2==2.9.6
django-CNPJField==1
django-cpf==0.1.0
django-crispy-forms==1.14.0
django-mptt==0.13.4
crispy-bootstrap5==0.6
djangorestframework==3.12.2
psycopg2-binary==2.9.3